Fixes thrust compilation issues.
