+++ src/mesodyn/density_initializer.h
@@ -23,63 +23,15 @@ class Molecule_density {
 
     public:
         // Used for solutes
-        Molecule_density(Molecule* molecule)
-        : Molecule_density(molecule, molecule->theta)
-        { }
-
+        Molecule_density(Molecule* molecule);
         // Used for solvents
-        Molecule_density(Molecule* molecule, Real molecule_total_mass)
-        : m_monomer_fraction_of_molecule(0),
-          m_molecule_total_mass{(size_t)molecule_total_mass},
-          m_total_monomer_densities(0),
-          m_lat(molecule->Lat[0])
-        {
-            assert(molecule->MolMonList.size() > 0);
-            // The indices in MolMonList are accepted by molecule->fraction
-            // and return the mass fraction of that particular monomer in the
-            // molecule
-            for(int& all_monomers : molecule->MolMonList)    
-                m_monomer_fraction_of_molecule.push_back(
-                    molecule->fraction(all_monomers)
-                );
+        Molecule_density(Molecule* molecule, Real molecule_total_mass);
 
-            set_total_monomer_densities();
-        }
-
-        void set_total_monomer_densities() {
-            for (Real& fraction : m_monomer_fraction_of_molecule) {
-                Real total_monomer_mass = m_molecule_total_mass * fraction;
-
-                m_total_monomer_densities.push_back(total_monomer_mass);
-            }
-        }
-
+        void set_total_monomer_densities();
         //Returns a vector of homogeneously distributed densities per monomer for this molecule
-        std::vector<Lattice_object<Real>> homogeneous(size_t system_volume) {
-            // argument system_volume: volume of the system minus the boundaries and solids
-            for (Real& total_monomer_mass : m_total_monomer_densities) {
-
-                // Masking objects is by in the mask class, which should probably be used after calling this function.
-                Real density_per_site = total_monomer_mass / system_volume;
-
-                Lattice_object<Real> this_monomer_density(m_lat, density_per_site);
-
-                m_homogeneous_monomer_densities.push_back(
-                    this_monomer_density
-                );
-
-            }
-          
-            return m_homogeneous_monomer_densities;
-        }
-
-        size_t total_mass() {
-            return m_molecule_total_mass;
-        }
-
-        vector<Real>& monomer_total_mass() {
-            return m_total_monomer_densities;
-        }
+        std::vector<Lattice_object<Real>> homogeneous(size_t system_volume);
+        size_t total_mass();
+        vector<Real>& monomer_total_mass();
 };
 
 //WARNING: THIS ASSUMES THAT SYSTEM HAS CALLED PREPAREFORCALCULATIONS
@@ -91,49 +43,9 @@ class Homogeneous_system_initializer {
         std::vector<Lattice_object<Real>> m_densities;
 
     public:
-        Homogeneous_system_initializer(System* system)
-        : SOLVENT{(size_t)system->solvent},
-          m_molecules{system->Mol},
-          m_system_volume{(size_t)system->boundaryless_volume}
-        { 
-            assert(system->boundaryless_volume > 0);
-            assert(system->Mol.size() > 1);
-            assert(system->Lat.size() > 0);
-        }
-
-        void build_objects()
-        {
-            std::vector< Molecule_density* > initializer(m_molecules.size());
-            Real total_solute_mass{0};
-
-            //We need to know total solute mass to calculate total solvent mass (theta) by subtracting
-            //Total solute mass from the total volume. Later, we initialize all densities at once so all
-            //densities per monomer stay in the same order as the input file.
-            for (size_t i = 0 ; i < m_molecules.size() ; ++i)
-                if (i != SOLVENT) {
-                    initializer[i] = new Molecule_density (
-                        m_molecules[i]);
-                    //Pool solute mass to find solvent mass
-                    total_solute_mass += initializer[i]->total_mass();
-                }
-
-            Real total_solvent_mass = m_system_volume - total_solute_mass;
-
-            //Initialize solvent with overloaded constructor
-            initializer[SOLVENT] = new Molecule_density (
-                    m_molecules[SOLVENT], total_solvent_mass);
-            
-            //Build rho in order of the input file.
-            for (Molecule_density* all_initializers : initializer)
-                for (auto all_densities : all_initializers->homogeneous(m_system_volume) )
-                    m_densities.push_back(all_densities);
-        }
-
-        void push_data_to_objects(std::vector<Lattice_object<Real>>& target) {
-            assert(target.size() == m_densities.size() && "Please resize your Lattice_object vector before passing!");
-            for (size_t i = 0 ; i < m_densities.size() ; ++i)
-                target[i] = m_densities[i];
-        }
+        Homogeneous_system_initializer(System* system);
+        void build_objects();
+        void push_data_to_objects(std::vector<Lattice_object<Real>>& target);
 };
 
 #endif
