+++ src/mesodyn/lattice_object.cpp
@@ -246,7 +246,7 @@ const stl::device_vector<T>& previous_state() {
 
 void reinstate_previous_state() {
   m_data = saved_state->state;
-  set_checkable_data_wrapper();
+  this->set_checkable_data( m_data.data(), system_size );
 }
 
 };
